refactor: update test cases for recent search use case.
