--- domain/src/test/java/com/example/domain/useCase/search/GetRecentSearchesUseCaseTest.kt
@@ -1,26 +1,0 @@
-package com.example.domain.useCase.search
-
-import com.example.domain.repository.RecentSearchRepository
-import io.mockk.coVerify
-import io.mockk.mockk
-import kotlinx.coroutines.runBlocking
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Test
-
-class GetRecentSearchesUseCaseTest {
-    private lateinit var recentSearchRepository: RecentSearchRepository
-    private lateinit var getRecentSearchesUseCase: GetRecentSearchesUseCase
-
-    @BeforeEach
-    fun setUp() {
-        recentSearchRepository = mockk(relaxed = true)
-        getRecentSearchesUseCase = GetRecentSearchesUseCase(recentSearchRepository)
-    }
-
-    @Test
-    fun `should call getAllRecentSearches when executed`() =
-        runBlocking {
-            getRecentSearchesUseCase()
-            coVerify { recentSearchRepository.getRecentSearches() }
-        }
-}
